requirements.txt
asgiref==3.4.1
Django==3.2.9
django-location-field==2.1.0
gunicorn==20.1.0
migrate==0.3.7
psycopg2==2.9.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
